Code CleanUps in Primitives / General (#350)
* EventHubCode Clean ups

* Code CleanUps in Primitives

* Rename variable

* Leave string cast

+++ src/Microsoft.Azure.EventHubs/Primitives/Ticks.cs
@@ -18,7 +18,7 @@ namespace Microsoft.Azure.EventHubs
 
         public static long FromMilliseconds(int milliseconds)
         {
-            return checked((long)milliseconds * TimeSpan.TicksPerMillisecond);
+            return checked(milliseconds * TimeSpan.TicksPerMillisecond);
         }
 
         public static int ToMilliseconds(long ticks)
